refactor(data-table): update story structure
