+++ src/components/DataTable/stories/shared.js
@@ -1,0 +1,61 @@
+import { action } from '@storybook/addon-actions';
+
+export const initialRows = [
+  {
+    id: 'a',
+    name: 'Load Balancer 3',
+    protocol: 'HTTP',
+    port: 3000,
+    rule: 'Round robin',
+    attached_groups: 'Kevins VM Groups',
+    status: 'Disabled',
+  },
+  {
+    id: 'b',
+    name: 'Load Balancer 1',
+    protocol: 'HTTP',
+    port: 443,
+    rule: 'Round robin',
+    attached_groups: 'Maureens VM Groups',
+    status: 'Starting',
+  },
+  {
+    id: 'c',
+    name: 'Load Balancer 2',
+    protocol: 'HTTP',
+    port: 80,
+    rule: 'DNS delegation',
+    attached_groups: 'Andrews VM Groups',
+    status: 'Active',
+  },
+];
+
+export const headers = [
+  {
+    key: 'name',
+    header: 'Name',
+  },
+  {
+    key: 'protocol',
+    header: 'Protocol',
+  },
+  {
+    key: 'port',
+    header: 'Port',
+  },
+  {
+    key: 'rule',
+    header: 'Rule',
+  },
+  {
+    key: 'attached_groups',
+    header: 'Attached Groups',
+  },
+  {
+    key: 'status',
+    header: 'Status',
+  },
+];
+
+export const batchActionClick = selectedRows => () =>
+  action('batch action click')(selectedRows);
