+++ src/components/DataTable/stories/with-selection.js
@@ -1,0 +1,46 @@
+import React from 'react';
+import DataTable, {
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableHeader,
+  TableRow,
+  TableSelectAll,
+  TableSelectRow,
+} from '../../DataTable';
+import { initialRows, headers } from './shared';
+
+export default () => (
+  <DataTable
+    rows={initialRows}
+    headers={headers}
+    render={({ rows, headers, getHeaderProps, getSelectionProps }) => (
+      <TableContainer title="DataTable">
+        <Table>
+          <TableHead>
+            <TableRow>
+              <TableSelectAll {...getSelectionProps()} />
+              {headers.map(header => (
+                <TableHeader {...getHeaderProps({ header })}>
+                  {header.header}
+                </TableHeader>
+              ))}
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {rows.map(row => (
+              <TableRow key={row.id}>
+                <TableSelectRow {...getSelectionProps({ row })} />
+                {row.cells.map(cell => (
+                  <TableCell key={cell.id}>{cell.value}</TableCell>
+                ))}
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </TableContainer>
+    )}
+  />
+);
