+++ src/components/DataTable/stories/with-sorting.js
@@ -1,0 +1,42 @@
+import React from 'react';
+import DataTable, {
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '../../DataTable';
+import { initialRows, headers } from './shared';
+
+export default () => (
+  <DataTable
+    rows={initialRows}
+    headers={headers}
+    render={({ rows, headers, getHeaderProps }) => (
+      <TableContainer title="DataTable with sorting">
+        <Table>
+          <TableHead>
+            <TableRow>
+              {headers.map(header => (
+                <TableHeader {...getHeaderProps({ header })}>
+                  {header.header}
+                </TableHeader>
+              ))}
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {rows.map(row => (
+              <TableRow key={row.id}>
+                {row.cells.map(cell => (
+                  <TableCell key={cell.id}>{cell.value}</TableCell>
+                ))}
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </TableContainer>
+    )}
+  />
+);
